+++ src/Fp.Fs/CHANGELOG.md
@@ -4,5 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+- Arg setup now happens in `Processor.Prepare(ProcessorConfiguration?)`
+- Format processor wrappers now run cleanup on the wrapped processors
+
 ## [0.25.0] 2022-1-18
 - Initial version